--- src/main/java/com/example/dailyboramspring/domain/profile/presentation/ProfileController.java
@@ -1,41 +1,0 @@
-package com.example.dailyboramspring.domain.profile.presentation;
-
-import com.example.dailyboramspring.domain.profile.presentation.dto.request.UpdateProfileRequest;
-import com.example.dailyboramspring.domain.profile.presentation.dto.response.MyUserProfileResponse;
-import com.example.dailyboramspring.domain.profile.presentation.dto.response.UserProfileResponse;
-import com.example.dailyboramspring.domain.profile.service.GetMyUserProfileService;
-import com.example.dailyboramspring.domain.profile.service.GetUserProfileService;
-import com.example.dailyboramspring.domain.profile.service.UpdateProfileService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-
-@RequiredArgsConstructor
-@RequestMapping("/profile")
-@RestController
-public class ProfileController {
-
-    private final GetUserProfileService getUserProfileService;
-    private final GetMyUserProfileService getMyUserProfileService;
-    private final UpdateProfileService updateProfileService;
-
-    @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/other/{user_id}")
-    public UserProfileResponse getUserProfile(@PathVariable(name = "user_id") Long id) {
-        return getUserProfileService.execute(id);
-    }
-
-    @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/me")
-    public MyUserProfileResponse getUserProfile() {
-        return getMyUserProfileService.execute();
-    }
-
-    @ResponseStatus(HttpStatus.OK)
-    @PatchMapping
-    public void updateProfile(@RequestBody @Valid UpdateProfileRequest updateProfileRequest) {
-        updateProfileService.execute(updateProfileRequest);
-    }
-}
