--- src/main/java/com/example/dailyboramspring/domain/profile/presentation/dto/response/PurchaseElement.java
@@ -1,15 +1,0 @@
-package com.example.dailyboramspring.domain.profile.presentation.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class PurchaseElement {
-
-    private String seriesTitle;
-
-    private String episodeTitle;
-
-    private Integer cost;
-}
