--- src/main/java/com/example/dailyboramspring/domain/profile/presentation/dto/response/SeriesElement.java
@@ -1,17 +1,0 @@
-package com.example.dailyboramspring.domain.profile.presentation.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SeriesElement {
-
-    private String title;
-
-    private String nickname;
-
-    private String image;
-
-    private Long like;
-}
