+++ src/main/java/com/example/dailyboramspring/domain/user/service/GetMyUserProfileService.java
@@ -1,15 +1,13 @@
 package com.example.dailyboramspring.domain.profile.service;
 
-import com.example.dailyboramspring.domain.profile.domain.Profile;
 import com.example.dailyboramspring.domain.profile.facade.ProfileFacade;
+import com.example.dailyboramspring.domain.profile.presentation.dto.response.MyUserProfileResponse;
+import com.example.dailyboramspring.domain.profile.presentation.dto.response.PurchaseElement;
 import com.example.dailyboramspring.domain.profile.presentation.dto.response.SeriesElement;
-import com.example.dailyboramspring.domain.profile.presentation.dto.response.UserProfileResponse;
-import com.example.dailyboramspring.domain.series.domain.Series;
+import com.example.dailyboramspring.domain.purchase.facade.PurchaseFacade;
 import com.example.dailyboramspring.domain.series.facade.SeriesFacade;
-import com.example.dailyboramspring.domain.serieslike.domain.SeriesLikeId;
 import com.example.dailyboramspring.domain.serieslike.facade.SeriesLikeFacade;
 import com.example.dailyboramspring.domain.user.domain.User;
-import com.example.dailyboramspring.domain.user.domain.repository.UserRepository;
 import com.example.dailyboramspring.domain.user.facade.UserFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,19 +15,21 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+
 @Service
 @RequiredArgsConstructor
-public class GetUserProfileService {
+public class GetMyUserProfileService {
 
     private final UserFacade userFacade;
     private final ProfileFacade profileFacade;
     private final SeriesFacade seriesFacade;
     private final SeriesLikeFacade seriesLikeFacade;
+    private final PurchaseFacade purchaseFacade;
 
 
-    public UserProfileResponse execute(Long id) {
+    public MyUserProfileResponse execute() {
 
-        User user = userFacade.getUserById(id);
+        User user = userFacade.getCurrentUser();
 
         Profile profile = profileFacade.getProfileById(user.getId());
 
@@ -43,11 +43,21 @@ public class GetUserProfileService {
                         .build())
                 .collect(Collectors.toList());
 
-        return new UserProfileResponse(
+        List<PurchaseElement> purchaseList = purchaseFacade.GetAllByUser(user)
+                .stream()
+                .map(purchase -> PurchaseElement.builder()
+                        .seriesTitle(purchase.getEpisode().getSeries().getTitle())
+                        .episodeTitle(purchase.getEpisode().getTitle())
+                        .cost(purchase.getEpisode().getCost())
+                        .build())
+                .collect(Collectors.toList());
+
+        return new MyUserProfileResponse(
                 profile.getNickname(),
                 profile.getImage(),
                 profile.getIntroduce(),
-                noticeList
+                noticeList,
+                purchaseList
         );
     }
 }
