+++ src/main/java/com/example/dailyboramspring/domain/user/service/GetUserProfileService.java
@@ -1,13 +1,8 @@
 package com.example.dailyboramspring.domain.profile.service;
 
-import com.example.dailyboramspring.domain.profile.domain.Profile;
 import com.example.dailyboramspring.domain.profile.facade.ProfileFacade;
-import com.example.dailyboramspring.domain.profile.presentation.dto.response.MyUserProfileResponse;
-import com.example.dailyboramspring.domain.profile.presentation.dto.response.PurchaseElement;
 import com.example.dailyboramspring.domain.profile.presentation.dto.response.SeriesElement;
 import com.example.dailyboramspring.domain.profile.presentation.dto.response.UserProfileResponse;
-import com.example.dailyboramspring.domain.purchase.facade.PurchaseFacade;
-import com.example.dailyboramspring.domain.series.domain.Series;
 import com.example.dailyboramspring.domain.series.facade.SeriesFacade;
 import com.example.dailyboramspring.domain.serieslike.facade.SeriesLikeFacade;
 import com.example.dailyboramspring.domain.user.domain.User;
@@ -18,21 +13,19 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
 @Service
 @RequiredArgsConstructor
-public class GetMyUserProfileService {
+public class GetUserProfileService {
 
     private final UserFacade userFacade;
     private final ProfileFacade profileFacade;
     private final SeriesFacade seriesFacade;
     private final SeriesLikeFacade seriesLikeFacade;
-    private final PurchaseFacade purchaseFacade;
 
 
-    public MyUserProfileResponse execute() {
+    public UserProfileResponse execute(Long id) {
 
-        User user = userFacade.getCurrentUser();
+        User user = userFacade.getUserById(id);
 
         Profile profile = profileFacade.getProfileById(user.getId());
 
@@ -46,21 +39,11 @@ public class GetMyUserProfileService {
                         .build())
                 .collect(Collectors.toList());
 
-        List<PurchaseElement> purchaseList = purchaseFacade.GetAllByUser(user)
-                .stream()
-                .map(purchase -> PurchaseElement.builder()
-                        .seriesTitle(purchase.getEpisode().getSeries().getTitle())
-                        .episodeTitle(purchase.getEpisode().getTitle())
-                        .cost(purchase.getEpisode().getCost())
-                        .build())
-                .collect(Collectors.toList());
-
-        return new MyUserProfileResponse(
+        return new UserProfileResponse(
                 profile.getNickname(),
                 profile.getImage(),
                 profile.getIntroduce(),
-                noticeList,
-                purchaseList
+                noticeList
         );
     }
 }
